+++ app/views/articles/show.html.haml
@@ -35,3 +35,5 @@
   = link_to new_article_comment_path(@article) do
     .btn-secondary
       コメントを追加
+
+= javascript_pack_tag 'article'
